backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
redis>=4.5
PyJWT>=2.6
django-cors-headers>=4.2
gunicorn>=21.2
